Vars Int to Float

+++ Assets/Scripts/StatusCharacters.cs
@@ -4,8 +4,8 @@ using System.Collections.Generic;
 [CreateAssetMenu(menuName = "CharacterCreator")]
 public class StatusCharacters : ScriptableObject
 {
-    public int hp;
-    public int damage;
+    public float hp;
+    public float damage;
     public bool AttackInTheTurn;
     public List<ISkill> skills;
 }
